#82 - update: change chest table

+++ core/application/src/main/java/org/naho/learning/result/ChestDetailResult.java
@@ -6,7 +6,7 @@ public record ChestDetailResult(
         Long id,
         ChestType chestType,
         String description,
-        Double minPoint,
-        Double maxPoint
+        Integer minPoint,
+        Integer maxPoint
 ) {
 }
